# server/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Admin System API",
    "version": "1.0.0",
    "description": "后台管理系统 API 文档 - 自动生成"
  },
  "servers": [
    {
      "url": "http://localhost:8080/"
    }
  ],
...
  "paths": {
...
    "/manage/sys/menu/getMenuTree": {
      "post": {
        "tags": [
          "菜单管理"
        ],
        "summary": "获取菜单树",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/manage/sys/menu/addMenuTree": {
      "post": {
        "tags": [
          "菜单管理"
        ],
        "summary": "添加菜单",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "menuName",
                  "menuCode",
                  "menuType"
                ],
                "properties": {
                  "parentId": {
                    "type": "string",
                    "description": "父菜单ID",
                    "example": null
                  },
                  "menuName": {
                    "type": "string",
                    "description": "菜单名称",
                    "example": "系统管理"
                  },
                  "menuCode": {
                    "type": "string",
                    "description": "菜单编码",
                    "example": "system"
                  },
                  "menuType": {
                    "type": "number",
                    "description": "菜单类型 1:目录 2:菜单 3:按钮",
                    "example": 1,
                    "enum": [
                      1,
                      2,
                      3
                    ]
                  },
                  "path": {
                    "type": "string",
                    "description": "路由路径",
                    "example": "/system"
                  },
                  "component": {
                    "type": "string",
                    "description": "组件路径",
                    "example": "Layout"
                  },
                  "icon": {
                    "type": "string",
                    "description": "图标",
                    "example": "SettingOutlined"
                  },
                  "sortOrder": {
                    "type": "number",
                    "description": "排序",
                    "example": 1
                  },
                  "visible": {
                    "type": "number",
                    "description": "是否可见 1:是 0:否",
                    "example": 1
                  },
                  "status": {
                    "type": "number",
                    "description": "状态 1:启用 0:禁用",
                    "example": 1
                  },
                  "permissionCode": {
                    "type": "string",
                    "description": "权限编码",
                    "example": "system:view"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/manage/sys/menu/editMenuTree": {
      "post": {
        "tags": [
          "菜单管理"
        ],
        "summary": "编辑菜单",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "id"
                ],
                "properties": {
                  "id": {
                    "type": "string",
                    "description": "菜单ID",
                    "example": "507f1f77bcf86cd799439011"
                  },
                  "parentId": {
                    "type": "string",
                    "description": "父菜单ID"
                  },
                  "menuName": {
                    "type": "string",
                    "description": "菜单名称"
                  },
                  "menuCode": {
                    "type": "string",
                    "description": "菜单编码"
                  },
                  "menuType": {
                    "type": "number",
                    "description": "菜单类型"
                  },
                  "path": {
                    "type": "string",
                    "description": "路由路径"
                  },
                  "component": {
                    "type": "string",
                    "description": "组件路径"
                  },
                  "icon": {
                    "type": "string",
                    "description": "图标"
                  },
                  "sortOrder": {
                    "type": "number",
                    "description": "排序"
                  },
                  "visible": {
                    "type": "number",
                    "description": "是否可见"
                  },
                  "status": {
                    "type": "number",
                    "description": "状态"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/manage/sys/menu/deleteMenuById": {
      "post": {
        "tags": [
          "菜单管理"
        ],
        "summary": "删除菜单",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "id"
                ],
                "properties": {
                  "id": {
                    "type": "string",
                    "description": "菜单ID",
                    "example": "507f1f77bcf86cd799439011"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/manage/sys/auth/getMenuTreeWithRole": {
      "post": {
        "tags": [
          "权限管理"
        ],
        "summary": "获取角色菜单权限树",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "roleId"
                ],
                "properties": {
                  "roleId": {
                    "type": "string",
                    "description": "角色ID",
                    "example": "507f1f77bcf86cd799439011"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/manage/sys/auth/getMenuTreeWithUser": {
      "post": {
        "tags": [
          "权限管理"
        ],
        "summary": "获取用户菜单权限树",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "userId"
                ],
                "properties": {
                  "userId": {
                    "type": "string",
                    "description": "用户ID",
                    "example": "507f1f77bcf86cd799439011"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/manage/sys/auth/saveRoleMenu": {
      "post": {
        "tags": [
          "权限管理"
        ],
        "summary": "保存角色菜单权限",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "roleId",
                  "menuIds"
                ],
                "properties": {
                  "roleId": {
                    "type": "string",
                    "description": "角色ID",
                    "example": "507f1f77bcf86cd799439011"
                  },
                  "menuIds": {
                    "type": "string",
                    "description": "菜单ID列表，逗号分隔",
                    "example": "id1,id2,id3"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
...
}